Build Completed Todo Rest API like visible

+++ src/main/java/net/javaguides/todo_management/service/impl/TodoServiceImpl.java
@@ -68,9 +68,20 @@ public class TodoServiceImpl implements TodoService {
 
     @Override
     public void deleteTodo(Long id) {
-       Todo todo = todoRepository.findById(id)
-                          .orElseThrow(()-> new ResourceNotFoundException("Todo not found with id" + id));
-//       todoRepository.delete(todo);
+        if (!todoRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Todo not found with id " + id);
+        }
        todoRepository.deleteById(id);
+    }
+
+    @Override
+    public TodoDto completeTodo(Long id) {
+        Todo todo = todoRepository.findById(id)
+                 .orElseThrow(()-> new ResourceNotFoundException("Todo not found with id " + id));
+        todo.setCompleted(true);
+
+        Todo updatedTodo = todoRepository.save(todo);
+
+        return modelMapper.map(updatedTodo, TodoDto.class);
     }
 }
